Refactor checker to enable single check, multiple profile compliance validation
