+++ clairmeta/dcp_check_isdcf_dcnc.py
@@ -10,8 +10,8 @@ from clairmeta.settings import DCP_SETTINGS
 
 
 class Checker(CheckerBase):
-    def __init__(self, dcp, profile):
-        super(Checker, self).__init__(dcp, profile)
+    def __init__(self, dcp):
+        super(Checker, self).__init__(dcp)
         self.fields = None
 
     def run_checks(self):
