+++ clairmeta/dcp_check_pkl.py
@@ -11,8 +11,8 @@ from clairmeta.dcp_utils import list_pkl_assets
 
 class Checker(CheckerBase):
 
-    def __init__(self, dcp, profile):
-        super(Checker, self).__init__(dcp, profile)
+    def __init__(self, dcp):
+        super(Checker, self).__init__(dcp)
 
     def run_checks(self):
         # Accumulate hash by UUID, useful for multi PKL package
