+++ clairmeta/dcp_check_vol.py
@@ -6,8 +6,8 @@ from clairmeta.dcp_check_utils import check_xml
 
 
 class Checker(CheckerBase):
-    def __init__(self, dcp, profile):
-        super(Checker, self).__init__(dcp, profile)
+    def __init__(self, dcp):
+        super(Checker, self).__init__(dcp)
 
     def run_checks(self):
         for source in self.dcp._list_vol:
